Remove more unneeded demo artifacts
